+++ back-end/src/products/products.controller.ts
@@ -6,6 +6,7 @@ import {
   Param,
   Patch,
   Delete,
+  BadRequestException
 } from '@nestjs/common';
 
 import { ProductsService } from './products.service';
@@ -21,6 +22,11 @@ export class ProductsController {
     @Body('price') prodPrice: number,
     @Body("image") prodImage: string 
   ) {
+    if(!prodTitle || !prodDesc || !prodPrice || !prodImage){
+      console.log(prodTitle,prodDesc,prodPrice,prodImage)
+      throw new BadRequestException('need more data');
+    }
+
     const generatedId = await this.productsService.insertProduct(
       prodTitle,
       prodDesc,
@@ -38,6 +44,9 @@ export class ProductsController {
 
   @Get(':id')
   getProduct(@Param('id') prodId: string) {
+    if (!prodId.match(/^[0-9a-fA-F]{24}$/)) {
+      throw new BadRequestException(`${prodId} is not a valid id`);
+    }
     return this.productsService.getSingleProduct(prodId);
   }
 
@@ -49,12 +58,22 @@ export class ProductsController {
     @Body('price') prodPrice: number,
     @Body('image') prodImage: string,
   ) {
+    if (!prodId.match(/^[0-9a-fA-F]{24}$/)) {
+      throw new BadRequestException(`${prodId} is not a valid id`);
+    }
+    
+    if(!prodTitle && !prodDesc && !prodPrice && !prodImage){
+      throw new BadRequestException('need more data');
+    }
     await this.productsService.updateProduct(prodId, prodTitle, prodDesc, prodPrice, prodImage);
     return null;
   }
 
   @Delete(':id')
   async removeProduct(@Param('id') prodId: string) {
+    if (!prodId.match(/^[0-9a-fA-F]{24}$/)) {
+      throw new BadRequestException(`${prodId} is not a valid id`);
+    }
       await this.productsService.deleteProduct(prodId);
       return null;
   }
